+++ app/[lng]/(auth)/(.)login/page.js
@@ -1,9 +1,36 @@
+'use client'
+
+import TagInput from '@/app/components/TagInput'
+
 import LoginForm from '../components/LoginForm'
 
-export default async function LoginPage() {
+const initialTags = [
+  'Adobe Photoshop',
+  'CorelDRAW',
+  'Canva',
+  'Adobe Lightroom',
+  'Figma',
+  'Adobe Illustrator',
+  'Sketch',
+  'DALL-E',
+  'Midjourney',
+  'Stable Diffusion',
+  'Leonardo.AI',
+]
+
+export default function LoginPage() {
+  const handleTagsChange = (tags) => {
+    console.log('Теги изменены:', tags)
+  }
+
   return (
     <>
       <h1 className="text-3xl text-red-900">Modal</h1>
+      <TagInput
+        initialTags={initialTags}
+        suggestionCount={5}
+        onTagsChange={handleTagsChange}
+      />
       <LoginForm />
     </>
   )
